utils/null_output_stream: move definitions into source file
